Add order management features with detail modal and list view
- Implemented OrderDetailModal for viewing order details, including user information, order details, subscription information, and webhook payload.
- Created OrdersList component for managing and filtering orders with search, status, and date range filters.
- Added tests for OrderDetailModal and OrdersList to ensure functionality and correctness.
- Included pagination and loading indicators in the orders list view.
- Enhanced user experience with action buttons for retrying provisioning and viewing details.

+++ app/Services/My8kApiClient.php
@@ -22,7 +22,7 @@ class My8kApiClient
     {
         $this->baseUrl = config('services.my8k.base_url', 'https://my8k.me/api/api.php');
         $this->apiKey = config('services.my8k.api_key');
-        $this->timeout = config('services.my8k.timeout', 30);
+        $this->timeout = (int) config('services.my8k.timeout', 30);
     }
 
     /**
